major code revision to prepare for second tracking algorithm
